+++ src/components/button/button.stories.js
@@ -1,17 +1,7 @@
 import "./button.css";
+import { createButton } from "./button";
 
 export default { title: "Button" };
-
-function createButton(buttonText) {
-  const button = document.createElement("button");
-  button.className = "btn";
-  const text = document.createTextNode(buttonText);
-  button.append(text);
-  button.addEventListener("click", () => {
-    alert("Thanks for shopping");
-  });
-  return button;
-}
 
 export const addToCart = () => {
   const button = createButton("Add to cart");
